pyproject.toml
[project]
name = "git-lfs-nextcloud-agent"
version = "0.1.0"
description = "Custom Git LFS transfer agent for Nextcloud via WebDAV"
authors = [
    {name = "no input"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
